Fix LD error for apps
There is some sort of bug in LD (??) that results in a linker error
_sometimes_ when using the `-r` argument (an assertion failure in
elflink.c:8300, for reference).

But it turns out we don't need two stages of LD, so we can just avoid
using `-r` completely. We had it because I was lazy when writing elf2bin
and didn't want to account for the case where there wasn't anything in
the GOT. I just fixed that...