"Downgrade" userland libc.a to thumbv6m
We were using thumbv7em which includes instructions not supported on
other ARM targets (notably division). Since thumbv7em is a strict
super-set of thumb7m, which is a strict super set of thumv6m, using the
lowest common denominator is fine (albeit inefficient).

The real solution will be to provide all options and have the build
system choose the right one.