Add LED capsule
Also change pin assignments in boards to separate LED pins from other
GPIO pins. Then update apps to use the new pin mappings, cleaning along
the way. A nice benefit of this approach is that we don't have to have a
header file that maps "LED_0" to the correct gpio pin number. To use the
first LED, just call the `led_*()` functions with the number 0.