add other interesting gcc warnings
I stole the old list of extra warnings from an old project that I'd
compiled several years ago now. This does a fresh pass to see what
else gcc has to offer now and adds a summary of all gcc warnings.

I strived to only turn on the ones that are useful and catch genuinely
interesting things. Some can be seen as a bit pedantic, specifically
the function prototype fellows that demand a (void) for functions
taking no parameters, but what you gain is the compiler identifying
functions that (c/sh)ould be marked static, which is great for preventing
symbol leakage an improving optimization, which I think is a worthwhile
tradeoff there.

Some of the attribute suggestion stuff may be annoying, and we can
revisit these as time goes on, but having built / cleaned up all of
the current userspace, I think this is a good set for now.