fix MPU region sizes
Debugging the MPU more revealed a Goldilocks situation: an off-by-one too large
and and off-by-one too small setting up MPU regions. Now we’re just right.
This explains two different issues we’ve been seeing:

  - From #289: Applications could write the region of memory directly below them
  - From @cxbrooks's exploration into very large apps that were seeing odd faults

First, an exploration of how the code was currently running:

Printing the MPU configuration:

    TOCK_DEBUG(313): num 0 start_addr 0x20006000 len 14 base 0x20006010 attr_and_size 0x300001d

In the case the goals is to allow the range 0x20006000-0x20008000

Decomposing the registers:

1. region_base_address is set to 0x20006010

    0010 0000 0000 0000 0110 0000 0001 0000
    |                     |          | |...-> region 0
    |                     |          |-----> valid = yes
    bit 31                bit 13

The top bits of this register are defined as

    /// [31:N]    | ADDR    | Region base address
    /// [(N-1):5] |         | Reserved

Which means we need N=13

2. attr_and_size is set to 0x300001d

The bottom are of interest here:

    /// 0     | ENABLE | Region enable
    /// 5:1   | SIZE   | Region size is 2^(SIZE+1) (minimum 3)

    0001 1101
      .. ...

    SIZE=01110, N=14

The effect is that the MPU sees a base address of 0x20004000, which
explains why it was allowing accesses lower than it should have.

The bad field is coming from the len parameter, which comes from

    let data_len = (32 - self.memory.len().leading_zeros()) as u32;

in contrast, the MPU documentation specifies that len should be set via

    ///   N = Log2(Region size in bytes)

And log2(0x2000) = 13, not 14. Hence our problem. Off-by-one too large.

Interestingly, the flash segment computed size slightly differently

    let text_len = ((32 - self.text.len().leading_zeros()) - 2) as u32;

This chose a region that was half as large as it should be. However, to
deal with alignment issues, we round up the reported size of applications,
which means that the top half of flash for applications was often unused
space, and this wouldn’t be an issue in practice. That is, until you start
flashing large apps.