be consistent about our defintion of `main` (#341)
Closes #338.

This mostly implements the discussion from the issue thread, it:

  - marks our `_start` in crt1 as weak
  - overrides `_start` in tests where access memory regions is nice
     - I view this as somewhat niche and a non-binding interface for now
  - defines that standard main prototype as `int main(void)`

In the original issue thread, we'd suggested a prototype of `void main(void)`,
as Tock currently ignores the return value of processes. Unfortuntely, `-Wall`
turns on `-Wmain` by default, which results in

    main.c:6:6: warning: return type of 'main' is not 'int' [-Wmain]
     void main(void) {
          ^~~~

while we could explicitly turn this off (`-Wno-main`), my current feeling is
that writing `int main` is pretty stock for C, and having a return value from a
process may be useful someday (i.e. returning 0 means this process meant to
exit). Sure, most things probably won't ever return, but I don't think it hurts
to stick with C convention here and let things return a value.