ADC capable of high-speed continuous sampling
Makes use of continuous ADC sampling on the SAM4L in order to provide an
ADC interface to userland capable of providing samples at a set
frequency. There are also two example applications `adc` and
`adc_continuous` which give examples of the synchronous and asynchronous
ADC interfaces provided to userland.

 * Specified ADC channels. Like GPIO, the ADC capsule is now initialized with a list of valid ADC channels.
 * Ability to continuously collect single samples at frequencies from 1 to 10000 Hz.
 * Ability to continuously collect buffers of samples at frequencies from 23 Hz to 187000 Hz.
 * Application interface for requesting a single buffer-full of samples at any frequency.
 * Application interface for requesting continuous double-buffered samples at any frequency.
 * Different method of initializing the ADC that _for some reason_ allows the hardware to work at arbitrary frequencies (the original method did not).