Call all expired userland alarms, not just first
If two alarms were expired (e.g. they have the same expiration value),
only the first called. This continues iterating through the alarm heap
until we reach an alarm that hasn't expired.

There are a couple subtelties to doing this, basically resulting from
the fact that we're calling the virtual alarm's callback directly:

  1. We need to re-read the current clock value each time we iterate
  through the heap since the alarm's callback could have taken a long
  time.

  2. The alarm's callback _could_ call `yield`, meaning that we would
  reenter the non-virtualized callback. So we need to take care not to
  overzealously dequeue virtual alarms from the heap before we know _for
  sure_ we're going handle them.