style: add format for libraries too
The way this is done right now has the slightly unfortunate effect
that `format_all` will reformat libtock for every application (for
the same reason that it "rebuilds" libtock (only it's already built)
when doing a `build_all`). This isn't a problem per se, it's just a
little slow.

I'll add that ergnomic nit to #432, but I think it's more valuable
to get consistent formatting in while I noodle over that problem