userland: warnings on use of certain functions
Inspired by a series of signpost bugs, there are some functions that,
while not inherently unsafe, are often dangerous to use and easy to
replace with something safer. The motivating example here being
`snprintf` in favor of `sprintf`. We could add more in the future*,
but for now this is a scaffolding

*Brad was lobbying for strlen, but that's a bit harder, critically
this warning only applies to _emitted_ function calls, so something
like `foo("buf", strlen("buf"))` would not emit a warning, but that's
a bridge for another day)