Remove data_start_pointer from load_info
data_start_pointer was only being used to set the value of r9 in
`switch_to_user`. However, first, that's dependant on the processes
(i.e. a processes that uses a different PIC strategy will point r9
somewhere else). Second, it was buggy for non-PIC processes (setting r9
to 0 in practice). Finally, we already have this value anyway from the
initial stack poitner in the default PIC strategy.

So, this removes that field, removes setting the r9 value in the kernel,
deferring it to the processes, and uses the initial stack pointer as the
reference for where the GOT starts.