gpio: decouple interrupt config and pin enable
Previously, there was no means for a caller to distinguish if a
syscall failed because of failure to enable the pin or failure to
configure the interrupt, which could leave the operation half-complete
opaquely to the user. As pin configuration is rare, just decouple
these.

As an added benefit, this brings the gpio and gpio_async interfaces
into alignment.