Move C++ exceptions table to end of ELF
We currently don't actually serialize it to the TBF (we probably should
be if we care about exceptions), so it was screwing up the computation
for where the rel.data section would be placed. Placing it at the end,
after the dummy `endsec` section, avoids that, and otherwise doesn't
make a difference.