Cleanup userland_generic.ld
Cleaned up a few of the comments, removed unused symbols and...
improtantly... made sure the text segment ends on a 4-byte aligned
boundary. RODATA can be non-word aligned (e.g. if there is an odd-sized
string), and if there's nothing to align it afterwards, it can end up
placing the RAM sections in non-word-aligned flash. On Cortex-M0 this
causes a fault when doing relocation since Cortex-M0 can only do
word-aligned memory loads.