WIP: native compilation
Most of the work here is making the build tooling flexible in the
arch family. This should be useful for things such as an eventual
Risc V port as well.

At this point, we're up through custom --entry, which is more of
a Mac-specific issue than a general native one. Taking a break here.