Fix mpu_walk_region_test
The memory actually made accessible by the kernel has chenged, so the
mpu_walk_region test is no longer correct. The previous version assumed
memory was accessible all the way to the grant region, when now it is
only (guaranteed to be) accessible until the memory break returned from
`sbrk(0)`.

This change fixes the test.