crt0 and .ld: Cleanup & add stack as section
The motivation for this change is to make sure that applications are
compiled with the GOT, data, and BSS sections at the relative offset in
memory that they will be when an app actually executes. Currently, these
sections are at an offset from the start of memory, when they are
actually used after the stack (and the stack is at the start of memory).

This works fine for PIC apps, since the GOT, data, and BSS sections can
be relocated both to where the app is in flash/memory, and to move them
past the stack. But, for apps compiled for fixed addresses, the GOT,
data, and BSS addresses cannot be arbitrarily moved, they must end up in
memory at the addresses the compiler expects them to be.

The key fix here is to add a `.stack` section to the linker file, so
that the GOT, data, and BSS sections are at the actual addresses they
will be at when the app actually executes. This has no real effect on
PIC apps, but allows fixed address apps to work. However, it does mean
that c_start needs to change to match the new offsets.

The main change to c_start is that now everything is relative to the
start of app flash and start of app memory, rather than the start of app
flash and the end of the ap stack. This mirrors the new linker script,
where everything is now at the correct address relative to the start of
app memory.

I also added comments everywhere since I was changing the lines of code
anyway, and it's not immediately obvious to me why everything works the
way it does.