Porting libtock-c to new 2.0 syscall API.
Replaces internal calls to relative and absolute alarms (command 4+5) in
reference+dt calls (command 6). This fixes bugs when applications can hang
because a very short alarm appears to the kernel as a very long one.

This also includes porting the LED system calls to the 2.0 API; without
this, the multi_alarm_test test application fails because it uses
the number of LEDs to determine how many alarms to create. Without this
fix it creates > 100 alarms and crashes immediately.