Port IPC to 2.0 driver
The only change in the IPC interface is the return type of `ipc_share`,
which is now an `allow_return_t` instead of an int. The implementation
of `ipc_discover` now performs a command in addition to the
allow_readonly to perform the discover and an additional allow to get
ownership of the buffer back.