"upcall" revers to a type of callback that is a direct function call passed to subscribe.
So the internal functions in temperature.c and humidity.c that
are passed to subscribe() are called upcalls, but arbitrary
function pointers passed from userpsace are called callbacks.
This way the API doesn't say whether it's an upcall or not,
but the internal implementation knows its own function is so
calls it such.