rv32i: fix yield system call class argument register
On RISC-V 32bit platforms, the system call class identifier for
yield (0) must be passed in the 5th argument register, that is a4 or
x14 respectively. Passing it in the 6th argument register (a5) causes
undefined behavior.

This bug was hard to trace down, because the system call tracing
output was misleading, putting the blame on other parts of the
codebase. The LiteX simualtor (Verilator) RISC-V formal interface,
together with the Ibex instruction traces helped to hunt this one
down.

Fixes: 7d91971d303564 ("RISC-V support for yield-no-wait.")
Signed-off-by: Leon Schuermann <leon@is.currently.online>