support: use deprecated instead of warning attribute
Clang support for the warning attribute has just landed but it isn't
compatible with this kind of usage, as it requires that the original
declaration be annotated with the same attribute. The warning attribute
is also meant not to trigger when the use of the warned about symbol
is eliminated through code optimizations, which isn't the point here.
We can use the deprecated attribute instead, solving these issues.