ipc: register service callback with service name
Use package name of service to register the service callback. This makes
it more clear API and makes the kernel side easier.

It will allow us to drop any special case code in kernel for IPC and it
will most likely allow us to drop the NUM_UPCALLS const generic as well
(and we can just use NUM_PROCS for both upcall and allows)

Change-Id: I86f272ba7815285fbbb89069647778c0388b0dc6