Configuration.mk: change litex_sim target definition to rv32imc
This is in response to issue tock/libtock-c#282, which exposed that
`rv32i` support is not prevalent among prebuilt RISC-V embedded
toolchains. Users of such toolchains are faced with linker errors when
trying to build a `rv32i` target, which is included by default in all
RISC-V builds. As of tock/tock#2860 the `litex_sim` board is actually
using a VexRiscv CPU with multiplication and compressed instruction
support and hence this target definition can be changed to target
`-march=rv32imc` instead. Yet, it does not solve the underlying
toolchain issue. Refer to [1] for more information.

Signed-off-by: Leon Schuermann <leon@is.currently.online>

[1]: https://github.com/tock/libtock-c/issues/282