replace warning_header.h with _FORTIFY_SOURCE
Fortify source basically does the same thing, only with static checks
of correctness (or runtime injection of checks) on more complete suite
of functions and with more internal compiler support than we can pull
off. Let's use the tool built for this job.

There were a bunch of things that didn't specify headers correctly
and got away with it becuase our build system was injecting the headers
implicitly. That was an unfortunate side effect of our approach...
This fixes that as well.

Closes #301.