example: i2c-usb-bridge: Initial support
This application bridges data to and from the serial console to the
I2C bus.
This allows a PC to communicate with an I2C device via a USB port.

On a PC you can connect to the UART of the device running this
application to access the I2C bus. The Tock device will then act as
a I2C master to interact with I2C devices.

Signed-off-by: Alistair Francis <alistair.francis@wdc.com>