Resolved mbedcrypto constructor failure
The somewhat long description will hopefully help catalog the reason for
this issue and expanation for the fix:

To simplify importing the openthread static library, the static
libraries were merged through a potentially flawed approach (extracting
the object files and rebundling into one archive). This approach had the
potential to cause name conflicts.

In testing, the OpenThread instance constructor would hang (due to a
subroutine of the constructor returning an error and triggering the
VerifyorExit() function to drop the program into an infinite loop). This
issue was traced to a failure when intitializing / configuring
mbedcrypto.

When originally compiling and linking the singular static library, I
experienced an error due to two heap related functions being undefined.
I resolved this by passing a CMake flag to openthread regarding the
heap usage in mbedcrypto (resolving the error). However, this was the
reason now for the constructor failing. This original compiling and
linking error was in fact due to a naming conflict due to extracting and
then merging all object files.

This was resolved by removing the aforementioned CMake flag and altering
the build system to keep each static library separate.