AppMakefile: depend on libc++ and newlib for cc,cpp,cxx build rules
This ensures that prior to building a C++ source file, both libc++ and
newlib are available. It fixes a race condition in the GitHub actions
CI which would fail by first building `examples/cxx_hello/main.cc`,
which proceeded to build it with neither newlib or libc++ fetched. We
make C++ objects depend on newlib as well, as they can feature
includes from the C standard library.