alarm: remove references to timezone
Use of timezone in `{get/set}timeofday` is deprecated.

If something did try to pass us the parameter, our implementation would
(incorrectly) ignore it. More significantly, our implementation does not
check that the `timeval` struct is nonnull before dereferencing it.

This updates libtock's interface to match our implementation and the
common-case scenario, and leaves the question of what to do if something
tries to use the legacy `tz` parameter to whomever implements the
wrapper.