make: libraries can depend on a version
Add `Makefile.version` so a library can express a particular file that
must exist for the library to be usable. If that file doesn't exist, the
build system knows to run Makefile.setup. This supports updating a
library, as even if the needed version changes the built artifacts will
still exist and make will not rebuild them.